Rebuilt tests using RSpec
As it turns out, RSpec does support nested describe blocks, but I was
attempting to put code directly into them, which is a no-no. Setting up
objects for testing must be done within an it block; instance variables
(@foo) set up in before blocks are out of scope inside describe until an
it block is encountered.

I like the output from RSpec better, and it's got a lot more support via
tools (TextMate, rcov) and documentation (The RSpec Book). I'll be
sticking with it for now, but I'm leaving the Shoulda tests in place;
they certainly aren't hurting anything.