Took another stab at making some view specs
Not working yet. Attempting to use datasets to define an assortment of
valid event pages, but not having much luck getting the view spec to
pull them in. I'm not sure this is a good way to isolate the view code,
anyhow, but it's all a learning experience.

At the very least, the datasets would make sense for more extensive
model testing. Maybe.