Added display of event-related boolean values
all_day_event and no_end_time are now check boxes on an event page, and
also columns in the pages table.

No logic yet to change behavior of the EventPage model or the Page view
according to these boolean values, but that's coming next. In the model:

  * all_day_event prevents display of times for the event, and all-day
    events get sorted before events with start times.
  * no_end_time prevents display of event_end time for the event, and it
    makes the event sort after all-day events, but before events with
    both start and end times.

And in the view:

  * all_day_event toggles display of event_start_row and event_end_row
    tr elements, as well as no_end_time check box.
  * no_end_time toggles display of event_end_row tr element.