Event date interface full of ajaxy goodness
Eureka! Found a third option without the negatives of the two options
mentioned in the previous checkin:

* Put remote_function into the onclick handlers of the all_day_event and
  no_end_time check boxes. Pro: Avoids unnecessary round-trips to the
  server; only clicks on the check boxes will fire an action. Con: Took
  some figuring out to get the form data to submit.

The solution to getting the form's data out was to add :submit =>
'edit_page' (edit_page being the id of the Page form), which correctly
submits all the field data from the form, with authenticity token.

Possible future enhancement: hook the ajax callbacks to display and hide
a throbber. It still takes a moment to ping the server and get it to
render partials, and the feedback would be nice. In practice, on a
production server, this probably won't be much of an issue, unless
you're logged in from Liberia over a 14.4 modem.