<r:events:each (from|to)='nonsense'> implemented, passes spec
As it turns out, Time.parse alone isn't enough to verify that something
is a proper date and time; it returns the current date and time if you
feed it 'wibble'. Time.parse relies on Date._parse (note underscore),
which has a really primitive and permissive parser; it returns an empty
hash {} for Date.parse('wibble') instead of throwing an exception.

Fixed the problem in event tags handling by making a call to Date.parse
(no underscore) before the Time.parse call. Date.parse is much stricter
than Date._parse (which is also, amusingly enough, undocumentd), and
when combined with Time.parse, should throw an ArgumentError if the
date string isn't really a date. I hope.